requirements.txt
requests==2.28.2
bs4~=0.0.1
beautifulsoup4==4.12.0
aiofiles~=23.1.0
python-slugify~=8.0.1
